Speaker/chasis modeling pass
* drop vendored biquad, use native miniaudio filters,
  adjust filter parameters for miniaudio
* CPU pushes raw square wave through a float buffer
  downstream, sound.c applies the filtering
* introduce chasis "boxiness" with ma_peak2 filter

+++ src/sound/sound.h
@@ -24,7 +24,7 @@
 
 int sound_init(struct em400_cfg_buzzer *cfg);
 void sound_shutdown(void);
-long sound_play(int16_t *buf, size_t frames);
+long sound_play(float *buf, size_t frames);
 void sound_start(void);
 void sound_stop(void);
